added 2 files - data.js, util.js, created module
